process github desktop workflow
renamed files testing theory about git handling of files with file names changes had earlier two versions of oneCherub under other file name, will I able to find that history? Is it clear by GitHub desktop icons (+-) that a filename change is captured as a deletion of the file and addition of a 'new file' and that the filename when unchanged stores a diff history for plain text but otherwise version history of nonbinary texts. Screenshot add of the process to remind us of this conversation on the BBoard.