BH: Copying original sketch to repo
I have quite a few Processing sketches that I'm planning on integrating
into MUDDLE, and to help other muddlers out I'm going to document
converting sketches into p5.js. There is already existing documentation
on how to do this, seen below:

https://github.com/processing/p5.js/wiki/Processing-transition

Hopefully seeing the process in action, with all of the potential
headaches, will be helpful to you. There is at least one caveat with
my strategy I can think of, which is that I'll be targeting modern
browsers for the sake of convenience. This means that some of the
JS syntax I'm using isn't equally supported with older browser versions,
but 90% browser coverage is pretty acceptable:

https://caniuse.com/#feat=es6-class