BH: Drop types from function signatures
As I described in the previous commit, JS infers types of things without
specifying them. In this case we do not need specify function argument
types or the return type of the function.

That also means that you have to be careful about attempting incorrect
actions on certain data types, like trying to take the square root of a
string.