BH: Refactor global variables
As mentioned before, Javascript does not let you specify variable types.
However, you can specify if the value of a variable will change during
during runtime using `let`, or if its immutable with `const`.

The `let` is similar to another keyword you might see in code examples
called `var`, but I suggest using the former since it's far more
descriptive. The `const` keyword is slightly misleading, because in most
languages `const` would mean that the value described is immutable.
That's true for `initBlackHoleSize` because its a number, but not true
of `system` because it's an Array. Arrays and Objects can have values
contained within them change, but because `system` has been described as
a constant Array it can only be an Array.