BH: Bugfixes, pt. 1
This is the issue with just building all of this out without testing...

Because `color()` is a function that comes to use from p5.js, and the
library is loaded dynamically, the library isn't immediately available
by the time our JS file runs. That means I had to put the color
definitions inside our `setup` function.