BH: Bugfixes, pt. 2
This is an issue with variable naming scope, similar to the last commit.
Because `color` is already the name of a p5.js function, I was locally
overriding it with a function argument called `color`.

Similarly `red`, `green`, and `blue` are all p5.js functions, but JS
assumed I was trying to use them before I finished instantiating them.
This is why you should always give names to variables/functions that
are context-free!