BH: The final bugfix!
Loading images in JS is a pain-in-the-ass, I opted not to bother. While
it's technically possible to do if we serve this out of Github or open
this site locally, it presents problems for hosting on Codepen.

Also `mousePressed` got renamed to `mouseIsPressed` between Processing
and p5.js.

And that's it! This works fine for me locally, just a little slower. I'm
sure this could be sped up since it's not particularly efficient, but
I'll leave this as-is for the sake of code legibility.